requirements.txt
# Web Framework
fastapi
uvicorn[standard]

# LLM and AI
google-generativeai
sentence-transformers

# Vector Database
faiss-cpu
numpy

# Database
pymongo

# Utilities
python-dotenv
python-docx
requests
PyPDF2
pydantic